14.6. Implementando serviço de cadastro de foto de produto

+++ gorisfood-api/src/main/java/br/com/felipe/gorisfood/domain/model/FotoProduto.java
@@ -27,5 +27,5 @@ public class FotoProduto {
 	private String nomeArquivo;
 	private String descricao;
 	private String contentType;
-	private Integer tamanho;
+	private Long tamanho;
 }
